docs: Comobobox の Story を分解
