+++ packages/smarthr-ui/src/components/ComboBox/SingleCombobox.stories.tsx
@@ -5,12 +5,11 @@ import React, { ReactNode, useCallback, useState } from 'react'
 import { FormControl } from '../FormControl'
 import { Stack } from '../Layout'
 
-import { MultiComboBox, SingleComboBox } from '.'
+import { SingleComboBox } from '.'
 
 export default {
-  title: 'Forms（フォーム）/ComboBox',
+  title: 'Forms（フォーム）/SingleComboBox',
   component: SingleComboBox,
-  subcomponents: { MultiComboBox },
   parameters: {
     docs: {
       source: { type: 'dynamic' },
@@ -70,7 +69,7 @@ const manyItems = Array.from({ length: 2000 }).map((_, i) => ({
 
 type Item = { label: ReactNode; value: string; disabled?: boolean; data?: any }
 
-export const Single: StoryFn = () => {
+export const SingleCombobox: StoryFn = () => {
   const [items, setItems] = useState<Item[]>(defaultItems)
   const [selectedItem, setSelectedItem] = useState<Item | null>(null)
   const [seq, setSeq] = useState(0)
@@ -293,223 +292,6 @@ export const Single: StoryFn = () => {
           />
         </FormControl>
       </form>
-    </Stack>
-  )
-}
-
-export const Multi: StoryFn = () => {
-  const [items, setItems] = useState<Item[]>(defaultItems)
-  const [selectedItems, setSelectedItems] = useState<Item[]>([])
-  const [seq, setSeq] = useState(0)
-  const [controlledInputValue, setControlledInputValue] = useState<string>('')
-
-  const handleSelectItem = useCallback(
-    (item: Item) => {
-      action('onSelect')(item)
-      setSelectedItems([...selectedItems, item])
-    },
-    [selectedItems],
-  )
-  const handleDelete = useCallback(
-    (deleted: Item) => {
-      action('onDelete')()
-      setSelectedItems(selectedItems.filter((item) => item.value !== deleted.value))
-    },
-    [selectedItems],
-  )
-  const handleAddItem = useCallback(
-    (label: string) => {
-      action('onAdd')(label)
-      const newItem = {
-        label,
-        value: label,
-      }
-      setItems([...items, newItem])
-      setSelectedItems([...selectedItems, newItem])
-      setSeq(seq + 1)
-    },
-    [items, selectedItems, seq],
-  )
-
-  return (
-    <Stack>
-      <FormControl title="デフォルト">
-        <MultiComboBox
-          name="default"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownHelpMessage="入力でフィルタリングできます。"
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-          onChangeSelected={(selected) => {
-            action('onChangeSelected')(selected)
-            setSelectedItems(selected)
-          }}
-          onFocus={action('onFocus')}
-          onBlur={action('onBlur')}
-          data-test="multi-combobox-default"
-        />
-      </FormControl>
-      <FormControl title="アイテム追加可能">
-        <MultiComboBox
-          name="onAdd"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownHelpMessage="新しいアイテムを追加できます。"
-          creatable
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-          onAdd={handleAddItem}
-          data-test="multi-combobox-creatable"
-        />
-      </FormControl>
-      <FormControl title="Disabled">
-        <MultiComboBox
-          name="disabled"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownHelpMessage="Disabled なコンボボックス"
-          disabled
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-          data-test="multi-combobox-disabled"
-        />
-      </FormControl>
-      <FormControl title="必須">
-        <MultiComboBox
-          name="required"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownHelpMessage="Required なコンボボックス"
-          required
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-          data-test="multi-combobox-disabled"
-        />
-      </FormControl>
-      <FormControl title="その他属性">
-        <MultiComboBox
-          name="inputAttributes"
-          items={items}
-          selectedItems={selectedItems}
-          inputAttributes={{
-            'aria-label': 'inputAttributes',
-          }}
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-          data-test="multi-combobox-disabled"
-        />
-      </FormControl>
-      <FormControl title="エラー表示">
-        <MultiComboBox
-          name="error"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownHelpMessage="入力でフィルタリングできます。"
-          error
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-        />
-      </FormControl>
-      <FormControl title="選択済みアイテムを省略表示 + ツールチップ">
-        <MultiComboBox
-          name="selectedItemEllipsis"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownHelpMessage="入力でフィルタリングできます。"
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-          selectedItemEllipsis
-        />
-      </FormControl>
-      <FormControl title="読込中">
-        <MultiComboBox
-          name="isLoading"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownHelpMessage="入力でフィルタリングできます。"
-          isLoading
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-        />
-      </FormControl>
-      <FormControl title="文言変更">
-        <MultiComboBox
-          name="decorator"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownHelpMessage="入力でフィルタリングできます"
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-          decorators={{
-            noResultText: (text) => `no result.(${text})`,
-            destroyButtonIconAlt: (text) => `destroy.(${text})`,
-            selectedListAriaLabel: (text) => `selected item list.(${text})`,
-          }}
-        />
-      </FormControl>
-      <FormControl title="選択解除ボタンを非表示">
-        <MultiComboBox
-          name="invisible_unselected_button"
-          items={items}
-          selectedItems={selectedItems.map((item) => ({ ...item, deletable: false }))}
-          dropdownHelpMessage="入力でフィルタリングできます。"
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-          data-test="multi-combobox-undeletable"
-        />
-      </FormControl>
-      <FormControl title="テキストボックスの挙動を制御">
-        <MultiComboBox
-          name="textbox_controllable"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownHelpMessage="入力でフィルタリングできます。"
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-          onChangeSelected={(selected) => {
-            action('onChangeSelected')(selected)
-            setSelectedItems(selected)
-          }}
-          inputValue={controlledInputValue}
-          onChangeInput={(e) => {
-            setControlledInputValue(e.target.value)
-          }}
-          onBlur={() => setControlledInputValue('')}
-        />
-      </FormControl>
-      <FormControl title="100%幅">
-        <MultiComboBox
-          name="widthFull"
-          items={items}
-          selectedItems={selectedItems}
-          width="100%"
-          dropdownHelpMessage="入力でフィルタリングできます。"
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-        />
-      </FormControl>
-      <FormControl title="ドロップダウンリストの幅を指定(38文字)">
-        <MultiComboBox
-          name="dropdownWidth38em"
-          items={items}
-          selectedItems={selectedItems}
-          dropdownWidth="38em"
-          dropdownHelpMessage="入力でフィルタリングできます"
-          onDelete={handleDelete}
-          onSelect={handleSelectItem}
-        />
-      </FormControl>
-      <FormControl title="アイテム数が多い時">
-        <MultiComboBox
-          name="manyItems"
-          items={manyItems}
-          selectedItems={[]}
-          dropdownHelpMessage="入力でフィルタリングできます。"
-          onSelect={action('onSelect')}
-          data-test="multi-combobox-many"
-        />
-      </FormControl>
     </Stack>
   )
 }
